server/requirements.txt
cryptography==45.0.4
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
langchain==0.3.26
langchain_google_genai==2.1.6
langchain_pinecone==0.2.8
pinecone==7.3.0
protobuf==6.31.1
PyJWT==2.9.0
python-dotenv==1.1.1
mysqlclient==2.1.1
django-cors-headers==4.1.0

